Updating with next generation code, still a bit of a wip
This is the current code used in the quickbbs disk caching code, and test_colors2 code.

Archives is mostly untouched, beyond some cleanup.

Cached Exists is a rebuild of directory_cached, but it has been built-out to include support for SHA224, SHA255 support as well as adding in extended attribute support